refactor(floating-monitor): collapse internal coordination
