+++ Android/src/app/src/main/java/com/ollitert/llm/server/ui/floatingmonitor/FloatingMonitorFormatter.kt
@@ -16,6 +16,8 @@
 
 package com.ollitert.llm.server.ui.floatingmonitor
 
+import com.ollitert.llm.server.common.ServerStatus
+
 private const val MAX_EXACT_COUNT = 99_999L
 private const val MAX_EXACT_ELAPSED_SECONDS = 999L
 private const val LAST_LATENCY_CAP_MS = 1_000_000L
@@ -30,6 +32,15 @@ data class FloatingMonitorLatencyText(
   val value: String,
   val unit: String?,
   val inlineUnit: String? = unit,
+)
+
+data class FloatingMonitorRenderModel(
+  val visualState: FloatingMonitorVisualState,
+  val requestValue: String,
+  val secondaryValue: String,
+  val secondaryLabel: String,
+  val secondaryUnit: String? = null,
+  val lastLatency: FloatingMonitorLatencyText? = null,
 )
 
 fun formatFloatingMonitorCount(count: Long): String {
@@ -68,3 +79,90 @@ fun formatLastLatency(latencyMs: Long): FloatingMonitorLatencyText {
   }
   return FloatingMonitorLatencyText(value = "$wholeSeconds.$tenths", unit = "s")
 }
+
+fun deriveFloatingMonitorRenderModel(
+  visualState: FloatingMonitorVisualState,
+  requestCount: Long,
+  errorCount: Long,
+  processingElapsedMillis: Long?,
+  lastLatencyMs: Long = 0L,
+): FloatingMonitorRenderModel? {
+  if (visualState == FloatingMonitorVisualState.Hidden) return null
+  val processingElapsed = formatProcessingElapsed(processingElapsedMillis ?: 0)
+
+  return FloatingMonitorRenderModel(
+    visualState = visualState,
+    requestValue = formatFloatingMonitorCount(requestCount),
+    secondaryValue = when (visualState) {
+      FloatingMonitorVisualState.Running -> formatFloatingMonitorCount(errorCount)
+      FloatingMonitorVisualState.Processing -> processingElapsed.value
+      FloatingMonitorVisualState.Hidden -> error("Hidden was handled above")
+    },
+    secondaryLabel = when (visualState) {
+      FloatingMonitorVisualState.Running -> "err"
+      FloatingMonitorVisualState.Processing -> "proc"
+      FloatingMonitorVisualState.Hidden -> error("Hidden was handled above")
+    },
+    secondaryUnit = when (visualState) {
+      FloatingMonitorVisualState.Running -> null
+      FloatingMonitorVisualState.Processing -> processingElapsed.inlineUnit
+      FloatingMonitorVisualState.Hidden -> error("Hidden was handled above")
+    },
+    lastLatency = when (visualState) {
+      FloatingMonitorVisualState.Running -> null
+      FloatingMonitorVisualState.Processing -> formatLastLatency(lastLatencyMs)
+      FloatingMonitorVisualState.Hidden -> error("Hidden was handled above")
+    },
+  )
+}
+
+fun floatingMonitorContentDescription(model: FloatingMonitorRenderModel): String =
+  when (model.visualState) {
+    FloatingMonitorVisualState.Running ->
+      "Running, requests ${model.requestValue}, errors ${model.secondaryValue}"
+    FloatingMonitorVisualState.Processing -> {
+      val lastLatency = model.lastLatency
+      val lastDescription = when (lastLatency?.unit) {
+        "ms" -> "${lastLatency.value} milliseconds"
+        "s" -> "${lastLatency.value} seconds"
+        else -> "unavailable"
+      }
+      "Processing, requests ${model.requestValue}, " +
+        "current processing ${model.secondaryValue} seconds, " +
+        "last successful latency $lastDescription"
+    }
+    FloatingMonitorVisualState.Hidden -> error("Hidden render models are not created")
+  }
+
+enum class FloatingMonitorVisualState {
+  Hidden,
+  Running,
+  Processing,
+}
+
+fun deriveFloatingMonitorVisualState(
+  status: ServerStatus,
+  isInferring: Boolean,
+): FloatingMonitorVisualState =
+  when {
+    status != ServerStatus.RUNNING -> FloatingMonitorVisualState.Hidden
+    isInferring -> FloatingMonitorVisualState.Processing
+    else -> FloatingMonitorVisualState.Running
+  }
+
+fun shouldShowFloatingMonitor(
+  settingEnabled: Boolean,
+  overlayPermissionGranted: Boolean,
+  permissionFlowInProgress: Boolean,
+  appIsForeground: Boolean,
+  launchSuppressionActive: Boolean,
+  serviceIsAlive: Boolean,
+  visualState: FloatingMonitorVisualState,
+): Boolean =
+  settingEnabled &&
+    overlayPermissionGranted &&
+    !permissionFlowInProgress &&
+    !appIsForeground &&
+    !launchSuppressionActive &&
+    serviceIsAlive &&
+    visualState != FloatingMonitorVisualState.Hidden
