--- Android/src/app/src/main/java/com/ollitert/llm/server/ui/floatingmonitor/FloatingMonitorState.kt
@@ -1,52 +1,0 @@
-/*
- * Copyright 2025-2026 @NightMean (https://github.com/NightMean)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ollitert.llm.server.ui.floatingmonitor
-
-import com.ollitert.llm.server.common.ServerStatus
-
-enum class FloatingMonitorVisualState {
-  Hidden,
-  Running,
-  Processing,
-}
-
-fun deriveFloatingMonitorVisualState(
-  status: ServerStatus,
-  isInferring: Boolean,
-): FloatingMonitorVisualState =
-  when {
-    status != ServerStatus.RUNNING -> FloatingMonitorVisualState.Hidden
-    isInferring -> FloatingMonitorVisualState.Processing
-    else -> FloatingMonitorVisualState.Running
-  }
-
-fun shouldShowFloatingMonitor(
-  settingEnabled: Boolean,
-  overlayPermissionGranted: Boolean,
-  permissionFlowInProgress: Boolean,
-  appIsForeground: Boolean,
-  launchSuppressionActive: Boolean,
-  serviceIsAlive: Boolean,
-  visualState: FloatingMonitorVisualState,
-): Boolean =
-  settingEnabled &&
-    overlayPermissionGranted &&
-    !permissionFlowInProgress &&
-    !appIsForeground &&
-    !launchSuppressionActive &&
-    serviceIsAlive &&
-    visualState != FloatingMonitorVisualState.Hidden
