--- Android/src/app/src/main/java/com/ollitert/llm/server/ui/floatingmonitor/FloatingMonitorTapSuppression.kt
@@ -1,66 +1,0 @@
-/*
- * Copyright 2025-2026 @NightMean (https://github.com/NightMean)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ollitert.llm.server.ui.floatingmonitor
-
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.launch
-
-internal class FloatingMonitorTapSuppression(
-  private val scope: CoroutineScope,
-  private val timeoutMillis: Long,
-  private val onReleased: () -> Unit = {},
-) {
-  private val _active = MutableStateFlow(false)
-  val active: StateFlow<Boolean> = _active.asStateFlow()
-
-  private var releaseJob: Job? = null
-
-  init {
-    require(timeoutMillis > 0L) { "timeoutMillis must be positive" }
-  }
-
-  fun suppress() {
-    releaseJob?.cancel()
-    _active.value = true
-    releaseJob = scope.launch {
-      delay(timeoutMillis)
-      if (!_active.value) return@launch
-      _active.value = false
-      releaseJob = null
-      onReleased()
-    }
-  }
-
-  fun clear() {
-    val wasActive = _active.value
-    releaseJob?.cancel()
-    releaseJob = null
-    _active.value = false
-    if (wasActive) onReleased()
-  }
-
-  fun dispose() {
-    releaseJob?.cancel()
-    releaseJob = null
-    _active.value = false
-  }
-}
