--- Android/src/app/src/main/java/com/ollitert/llm/server/ui/floatingmonitor/ProcessingElapsedTracker.kt
@@ -1,45 +1,0 @@
-/*
- * Copyright 2025-2026 @NightMean (https://github.com/NightMean)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ollitert.llm.server.ui.floatingmonitor
-
-class ProcessingElapsedTracker(
-  private val elapsedRealtime: () -> Long,
-) {
-  private var processingStartedAt: Long? = null
-  private var activeInferenceSequence: Long? = null
-
-  fun update(isInferring: Boolean, inferenceSequence: Long) {
-    if (isInferring) {
-      if (activeInferenceSequence != inferenceSequence) {
-        activeInferenceSequence = inferenceSequence
-        processingStartedAt = elapsedRealtime()
-      }
-    } else {
-      activeInferenceSequence = null
-      processingStartedAt = null
-    }
-  }
-
-  fun elapsedMillis(): Long? = processingStartedAt?.let { startedAt ->
-    (elapsedRealtime() - startedAt).coerceAtLeast(0)
-  }
-
-  fun dispose() {
-    activeInferenceSequence = null
-    processingStartedAt = null
-  }
-}
